+++ bindings/java/iota-wallet-java/lib/src/main/java/org/iota/api/NativeApi.java
@@ -9,6 +9,9 @@ import org.apache.commons.lang3.SystemUtils;
 import org.iota.types.WalletConfig;
 import org.iota.types.exceptions.InitializeWalletException;
 import org.iota.types.exceptions.WalletException;
+import org.iota.types.events.Event;
+import org.iota.types.events.EventListener;
+import org.iota.types.events.wallet.WalletEventType;
 
 public class NativeApi {
 
@@ -23,7 +26,7 @@ public class NativeApi {
             loadFromJavaPathThrowable = t;
         }
 
-        if(loadFromJavaPathThrowable != null) {
+        if (loadFromJavaPathThrowable != null) {
             try {
                 loadFromJar();
             } catch (Throwable t) {
@@ -31,7 +34,7 @@ public class NativeApi {
             }
         }
 
-        if(loadFromJavaPathThrowable != null && loadFromJarThrowable != null) {
+        if (loadFromJavaPathThrowable != null && loadFromJarThrowable != null) {
             loadFromJavaPathThrowable.printStackTrace();
             loadFromJarThrowable.printStackTrace();
             throw new RuntimeException("cannot load native library");
@@ -52,7 +55,8 @@ public class NativeApi {
             libraryName = "libiota_wallet.dylib";
         else if (SystemUtils.IS_OS_WINDOWS)
             libraryName = "iota_wallet.dll";
-        else throw new RuntimeException("OS not supported");
+        else
+            throw new RuntimeException("OS not supported");
 
         NativeUtils.loadLibraryFromJar("/" + libraryName);
     }
@@ -65,25 +69,55 @@ public class NativeApi {
         }
     }
 
+    protected static native void initLogger(String config);
+
     private static native void createMessageHandler(String config) throws Exception;
+
+    // Destroys account handle
+    protected static native void destroyHandle();
 
     private static native String sendMessage(String command);
 
-    public static JsonElement callBaseApi(WalletCommand command) throws WalletException {
-        //System.out.println("REQUEST: " + command);
-        String jsonResponse = sendMessage(command.toString());
-        //System.out.println("RESPONSE: " + jsonResponse);
+    private static native String listen(String[] events, EventListener listener);
+
+    private static JsonElement handleClientResponse(String methodName, String jsonResponse) throws WalletException {
         WalletResponse response = CustomGson.get().fromJson(jsonResponse, WalletResponse.class);
 
         switch (response.type) {
             case "panic":
                 throw new RuntimeException(response.toString());
             case "error":
-                throw new WalletException(command.getMethodName(), response.payload.getAsJsonObject().toString());
+                throw new WalletException(methodName, response.payload.getAsJsonObject().toString());
 
             default:
                 return response.payload;
         }
+    }
+
+    private static void handleCallback(String response, EventListener listener) throws WalletException {
+        try {
+            Event event = CustomGson.get().fromJson(response, Event.class);
+            listener.receive(event);
+        } catch (Exception e) {
+            throw new WalletException("handleCallback", e.getMessage());
+        }
+    }
+
+    public static void callListen(EventListener listener, WalletEventType... events) throws WalletException {
+        String[] eventStrs = new String[events.length];
+        for (int i = 0; i < events.length; i++) {
+            eventStrs[i] = events[i].toString();
+        }
+
+        // Check for errors, no interest in result
+        handleClientResponse("listen", listen(eventStrs, listener));
+    }
+
+    public static JsonElement callBaseApi(WalletCommand command) throws WalletException {
+        // System.out.println("REQUEST: " + command);
+        String jsonResponse = sendMessage(command.toString());
+        // System.out.println("RESPONSE: " + jsonResponse);
+        return handleClientResponse(command.getMethodName(), jsonResponse);
     }
 
     private class WalletResponse {