+++ bindings/java/iota-wallet-java/lib/src/main/java/org/iota/types/events/EventListener.java
@@ -1,0 +1,10 @@
+package org.iota.types.events;
+
+public interface EventListener {
+    /**
+     * Event callback triggered at certain wallet actions
+     * 
+     * @param event The event
+     */
+    public void receive(Event event);
+}