+++ bindings/java/iota-wallet-java/lib/src/main/java/org/iota/types/events/wallet/ConsolidationRequired.java
@@ -1,0 +1,10 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+package org.iota.types.events.wallet;
+
+public class ConsolidationRequired extends WalletEvent {
+    public ConsolidationRequired() {
+        super(WalletEventType.ConsolidationRequired);
+    }
+}