+++ src/message_interface/message.rs
@@ -13,8 +13,7 @@ use serde::{Deserialize, Serialize};
 
 use super::account_method::AccountMethod;
 #[cfg(feature = "events")]
-#[cfg(debug_assertions)]
-use crate::events::types::WalletEvent;
+use crate::events::types::{WalletEvent, WalletEventType};
 use crate::{
     account::{operations::syncing::SyncOptions, types::AccountIdentifier},
     ClientOptions,
@@ -154,7 +153,6 @@ pub enum Message {
     /// Emits an event for testing if the event system is working
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     #[cfg(feature = "events")]
-    #[cfg(debug_assertions)]
     EmitTestEvent(WalletEvent),
     /// Transforms bech32 to hex
     /// Expected response: [`HexAddress`](crate::message_interface::Response::HexAddress)
@@ -168,6 +166,10 @@ pub enum Message {
         #[serde(rename = "bech32Hrp")]
         bech32_hrp: Option<String>,
     },
+    // Remove all listeners of this type. Empty vec clears all listeners
+    /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
+    #[cfg(feature = "events")]
+    ClearListeners(Vec<WalletEventType>),
     /// Stores participation information locally and returns the event.
     ///
     /// This will NOT store the node url and auth inside the client options.
@@ -253,12 +255,14 @@ impl Debug for Message {
             ),
             Message::StopBackgroundSync => write!(f, "StopBackgroundSync"),
             #[cfg(feature = "events")]
-            #[cfg(debug_assertions)]
             Message::EmitTestEvent(event) => write!(f, "EmitTestEvent({:?})", event),
             Message::Bech32ToHex(bech32_address) => write!(f, "Bech32ToHex({:?})", bech32_address),
             Message::HexToBech32 { hex, bech32_hrp } => {
                 write!(f, "HexToBech32{{ hex: {:?}, bech32_hrp: {:?} }}", hex, bech32_hrp)
             }
+
+            #[cfg(feature = "events")]
+            Message::ClearListeners(events) => write!(f, "ClearListeners({:?})", events),
             #[cfg(feature = "participation")]
             Message::RegisterParticipationEvent { event_id, nodes } => {
                 write!(